+++ pkg/compactor/blocks_cleaner.go
@@ -419,9 +419,23 @@ func (c *BlocksCleaner) cleanUserPartialBlocks(ctx context.Context, partials map
 		// We can safely delete only partial blocks with a deletion mark.
 		err := metadata.ReadMarker(ctx, userLogger, userBucket, blockID.String(), &metadata.DeletionMark{})
 		if errors.Is(err, metadata.ErrorMarkerNotFound) {
-			return nil
-		}
-		if err != nil {
+			//If only visit marker exists in the block, we can safely delete it.
+			isEmpty := true
+			notVisitMarkerError := userBucket.ReaderWithExpectedErrs(IsNotBlockVisitMarkerError).Iter(ctx, blockID.String(), func(file string) error {
+				isEmpty = false
+				if !IsBlockVisitMarker(file) {
+					// return error here to fail iteration fast
+					// to avoid going through all files
+					return ErrorNotBlockVisitMarker
+				}
+				return nil
+			})
+			if isEmpty || notVisitMarkerError != nil {
+				// skip deleting partial block if block directory
+				// is empty or non visit marker file exists
+				return nil
+			}
+		} else if err != nil {
 			level.Warn(userLogger).Log("msg", "error reading partial block deletion mark", "block", blockID, "err", err)
 			return nil
 		}