+++ pkg/compactor/compactor_test.go
@@ -1423,6 +1423,14 @@ func createNoCompactionMark(t *testing.T, bkt objstore.Bucket, userID string, bl
 	require.NoError(t, bkt.Upload(context.Background(), markPath, strings.NewReader(content)))
 }
 
+func createBlockVisitMarker(t *testing.T, bkt objstore.Bucket, userID string, blockID ulid.ULID) {
+	content := mockBlockVisitMarker()
+	blockPath := path.Join(userID, blockID.String())
+	markPath := path.Join(blockPath, BlockVisitMarkerFile)
+
+	require.NoError(t, bkt.Upload(context.Background(), markPath, strings.NewReader(content)))
+}
+
 func findCompactorByUserID(compactors []*Compactor, logs []*concurrency.SyncBuffer, userID string) (*Compactor, *concurrency.SyncBuffer, error) {
 	var compactor *Compactor
 	var log *concurrency.SyncBuffer
@@ -1675,6 +1683,21 @@ func mockBlockMetaJSONWithTime(id string, orgID string, minTime int64, maxTime i
 	content, err := json.Marshal(meta)
 	if err != nil {
 		panic("failed to marshal mocked block meta")
+	}
+
+	return string(content)
+}
+
+func mockBlockVisitMarker() string {
+	blockVisitMarker := BlockVisitMarker{
+		CompactorID: "dummy",
+		VisitTime:   time.Now().Unix(),
+		Version:     1,
+	}
+
+	content, err := json.Marshal(blockVisitMarker)
+	if err != nil {
+		panic("failed to marshal mocked block visit marker")
 	}
 
 	return string(content)